Split adapters for easier management
