+++ core/adapters/__init__.py
@@ -1,0 +1,14 @@
+"""core.adapters: This module defines adapters for the ACE application, providing
+a layer of abstraction for different APIs the tools might depend on."""
+
+from .services import RequestsHTTPAdapter
+from .storage import DatabaseConversationStorageAdapter, DatabaseLogStorageAdapter
+from .views import BuiltinIOAdapter, RichIOAdapter
+
+__all__ = [
+    "BuiltinIOAdapter",
+    "RichIOAdapter",
+    "RequestsHTTPAdapter",
+    "DatabaseConversationStorageAdapter",
+    "DatabaseLogStorageAdapter",
+]
