+++ tests/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""tests.adapters: Ensure that the adapters are correctly implemented and can be called as expected."""
